Event-Driven_Programming_Routing
Routing refers to the mechanism for serving the client content it has asked for. 
For web-based client/server applications, the client specifies the desired content in the URL; specifically, the path and querystring.
Server routing traditionally hinges on the path and the querystring, but there is other information available: headers, the domain, the IP address. This allow servers to take into consideration, for example, the approximate physical location of the user or preferred language of the user.